--- src/main/java/com/bcc/washer/controller/PublicController.java
@@ -1,14 +1,0 @@
-package com.bcc.washer.controller;
-
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
-
-@RestController
-@RequestMapping("/public")
-public class PublicController {
-
-    @GetMapping("/test")
-    public ResponseEntity<String> test() {
-        return ResponseEntity.ok("Public endpoint working");
-    }
-}
