+++ src/main/java/com/bcc/washer/domain/BookableUnit.java
@@ -1,5 +1,8 @@
 package com.bcc.washer.domain;
 
+import com.bcc.washer.domain.reservation.Reservation;
+import com.bcc.washer.domain.time.TimeSlot;
+import com.bcc.washer.domain.washer.Washer;
 import jakarta.persistence.*;
 import lombok.*;
