--- src/main/java/com/bcc/washer/domain/TimeSlot.java
@@ -1,29 +1,0 @@
-package com.bcc.washer.domain;
-
-import com.fasterxml.jackson.annotation.JsonIgnore;
-import jakarta.persistence.*;
-import lombok.*;
-
-import java.util.List;
-
-
-@Entity
-@Builder
-@Setter
-@Getter
-@AllArgsConstructor
-@NoArgsConstructor
-public class TimeSlot {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY )
-    private Long id;
-
-    @OneToOne( cascade = CascadeType.ALL, orphanRemoval = true)
-    private TimeInterval timeInterval;
-
-    @JsonIgnore
-    @OneToMany(mappedBy = "timeSlot", cascade = CascadeType.ALL, orphanRemoval = true)
-    private List<BookableUnit> bookableUnit;
-
-}
