+++ src/main/java/com/bcc/washer/domain/washer/WasherStatus.java
@@ -1,0 +1,7 @@
+
+package com.bcc.washer.domain.washer;
+
+public enum WasherStatus {
+    AVAILABLE,
+    MAINTENANCE
+}
