+++ src/main/java/com/bcc/washer/repository/TimeIntervalRepository.java
@@ -1,9 +1,7 @@
 package com.bcc.washer.repository;
 
-import com.bcc.washer.domain.TimeInterval;
+import com.bcc.washer.domain.time.TimeInterval;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
 
-@Repository
-public interface TimeIntervalRepository  extends JpaRepository<TimeInterval, Long> {
+public interface TimeIntervalRepository extends JpaRepository<TimeInterval, Long> {
 }
