+++ src/main/java/com/bcc/washer/repository/TimeSlotRepository.java
@@ -1,6 +1,6 @@
 package com.bcc.washer.repository;
 
-import com.bcc.washer.domain.TimeSlot;
+import com.bcc.washer.domain.time.TimeSlot;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
