+++ src/main/java/com/bcc/washer/repository/UserRepository.java
@@ -1,6 +1,6 @@
 package com.bcc.washer.repository;
 
-import com.bcc.washer.domain.User;
+import com.bcc.washer.domain.user.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
